+++ relm4/src/extensions/container.rs
@@ -28,6 +28,13 @@ impl RelmContainerExt for adw::PreferencesGroup {
     }
 }
 
+#[cfg(feature = "libadwaita")]
+impl RelmContainerExt for adw::Squeezer {
+    fn container_add(&self, widget: &impl AsRef<gtk::Widget>) {
+        self.add(widget.as_ref());
+    }
+}
+
 macro_rules! append_impl {
     ($($type:ty),+) => {
         $(