+++ CHANGES.md
@@ -5,6 +5,8 @@
 ### Added
 
 + core: Add `gnome_42` and `gnome_43` feature flags
++ core: Implement `RelmContainerExt` for `adw::Squeezer`
++ core: Implement `RelmSetChildExt` for `gtk::WindowHandle`
 
 ### Fixed